Bug 1610512 - Part 1: Update in-tree ICU to release 66.1. r=jwalden
Differential Revision: https://phabricator.services.mozilla.com/D66555

UltraBlame original commit: ffa621c545e10fa997b82ff13d4405fd3a351685

+++ intl/icu/source/common/ucnv.cpp
@@ -234,7 +234,7 @@ ucnv_safeClone(const UConverter* cnv, void *stackBuffer, int32_t *pBufferSize, U
         ptrdiff_t pointerAdjustment = aligned_p - p;
         if (bufferSizeNeeded + pointerAdjustment <= stackBufferSize) {
             stackBuffer = reinterpret_cast<void *>(aligned_p);
-            stackBufferSize -= pointerAdjustment;
+            stackBufferSize -= static_cast<int32_t>(pointerAdjustment);
         } else {
             
             stackBufferSize = 1;
